+++ F1/F1/Driver/DriverViewController.swift
@@ -98,5 +98,6 @@ extension  DriverViewController: ViewModelDelegate {
     }
 
     func show(error: String) {
+        showAlert(alertTitle: "Error", alertMessage: "Oops, an error occurred")
     }
 }
